Batch serialization was modified to explicitly include the signature size of the requests. The same was done for the encoding and decoding of messages exchanged via the netty framework between clients and replicas (in this case not only for signatures, but for MACs and request payload too).
This was done because previous commits included changes to allow developers to supply their own cryptographic providers and public key loaders, which means that now the library may have to use algorithms that to not enforce a fixed size on signatures (e.g., ECDSA signatures). Therefore it is not be safe anymore to implicitly assume the size of the signatures during deserialization.

This modification caused serialized batches to gain an extra 4 bytes for each signed request, and between 8 to 12 additional bytes to serialized messages exchanged with Netty. Therefore, a slight throughput decrease may be observed when submitting the library to new workload tests.